+++ app/ante/ante_test.go
@@ -1,9 +1,11 @@
 package ante_test
 
 import (
-	"github.com/cosmos/cosmos-sdk/types/tx/signing"
+	"errors"
 	"math/big"
 	"strings"
+
+	"github.com/cosmos/cosmos-sdk/types/tx/signing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -684,4 +686,137 @@ func (suite AnteTestSuite) TestAnteHandlerWithDynamicTxFee() {
 	}
 	suite.enableFeemarket = false
 	suite.enableLondonHF = true
+}
+
+func (suite AnteTestSuite) TestAnteHandlerWithParams() {
+	addr, privKey := tests.NewAddrKey()
+	to := tests.GenerateAddress()
+
+	testCases := []struct {
+		name         string
+		txFn         func() sdk.Tx
+		enableCall   bool
+		enableCreate bool
+		expErr       error
+	}{
+		{
+			"fail - Contract Creation Disabled",
+			func() sdk.Tx {
+				signedContractTx :=
+					evmtypes.NewTxContract(
+						suite.app.EvmKeeper.ChainID(),
+						1,
+						big.NewInt(10),
+						100000,
+						nil,
+						big.NewInt(ethparams.InitialBaseFee+1),
+						big.NewInt(1),
+						nil,
+						&types.AccessList{},
+					)
+				signedContractTx.From = addr.Hex()
+
+				tx := suite.CreateTestTx(signedContractTx, privKey, 1, false)
+				return tx
+			},
+			true, false,
+			evmtypes.ErrCreateDisabled,
+		},
+		{
+			"success - Contract Creation Enabled",
+			func() sdk.Tx {
+				signedContractTx :=
+					evmtypes.NewTxContract(
+						suite.app.EvmKeeper.ChainID(),
+						1,
+						big.NewInt(10),
+						100000,
+						nil,
+						big.NewInt(ethparams.InitialBaseFee+1),
+						big.NewInt(1),
+						nil,
+						&types.AccessList{},
+					)
+				signedContractTx.From = addr.Hex()
+
+				tx := suite.CreateTestTx(signedContractTx, privKey, 1, false)
+				return tx
+			},
+			true, true,
+			nil,
+		},
+		{
+			"fail - EVM Call Disabled",
+			func() sdk.Tx {
+				signedTx :=
+					evmtypes.NewTx(
+						suite.app.EvmKeeper.ChainID(),
+						1,
+						&to,
+						big.NewInt(10),
+						100000,
+						nil,
+						big.NewInt(ethparams.InitialBaseFee+1),
+						big.NewInt(1),
+						nil,
+						&types.AccessList{},
+					)
+				signedTx.From = addr.Hex()
+
+				tx := suite.CreateTestTx(signedTx, privKey, 1, false)
+				return tx
+			},
+			false, true,
+			evmtypes.ErrCallDisabled,
+		},
+		{
+			"success - EVM Call Enabled",
+			func() sdk.Tx {
+				signedTx :=
+					evmtypes.NewTx(
+						suite.app.EvmKeeper.ChainID(),
+						1,
+						&to,
+						big.NewInt(10),
+						100000,
+						nil,
+						big.NewInt(ethparams.InitialBaseFee+1),
+						big.NewInt(1),
+						nil,
+						&types.AccessList{},
+					)
+				signedTx.From = addr.Hex()
+
+				tx := suite.CreateTestTx(signedTx, privKey, 1, false)
+				return tx
+			},
+			true, true,
+			nil,
+		},
+	}
+
+	for _, tc := range testCases {
+		suite.Run(tc.name, func() {
+			suite.evmParamsOption = func(params *evmtypes.Params) {
+				params.EnableCall = tc.enableCall
+				params.EnableCreate = tc.enableCreate
+			}
+			suite.SetupTest() // reset
+
+			acc := suite.app.AccountKeeper.NewAccountWithAddress(suite.ctx, addr.Bytes())
+			suite.Require().NoError(acc.SetSequence(1))
+			suite.app.AccountKeeper.SetAccount(suite.ctx, acc)
+
+			suite.ctx = suite.ctx.WithIsCheckTx(true)
+			suite.app.EvmKeeper.SetBalance(suite.ctx, addr, big.NewInt((ethparams.InitialBaseFee+10)*100000))
+			_, err := suite.anteHandler(suite.ctx, tc.txFn(), false)
+			if tc.expErr == nil {
+				suite.Require().NoError(err)
+			} else {
+				suite.Require().Error(err)
+				suite.Require().True(errors.Is(err, tc.expErr))
+			}
+		})
+	}
+	suite.evmParamsOption = nil
 }